# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5fee94e5a3e0491123239459
  title: Educom
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /class:
    post:
      x-skaffolder-id: 5ff0adf1a3e049112323987c
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5fee9a81a3e04911232394d1
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Class
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                classCode:
                  type: string
                className:
                  type: string
    get:
      x-skaffolder-id: 5ff0ae16a3e0491123239883
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5fee9a81a3e04911232394d1
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Class
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
  "/class/{id}":
    delete:
      x-skaffolder-id: 5ff0ae16a3e0491123239882
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5fee9a81a3e04911232394d1
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Class
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ff0adf1a3e049112323987d
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5fee9a81a3e04911232394d1
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Class
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    post:
      x-skaffolder-id: 5ff0adf1a3e049112323987b
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5fee9a81a3e04911232394d1
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Class
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Class"
  "/class/findByclassSubject/{key}":
    get:
      x-skaffolder-id: 5ff0aea3a3e0491123239890
      x-skaffolder-name: findByclassSubject
      x-skaffolder-id-resource: 5fee9a81a3e04911232394d1
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: findByclassSubject
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByclassSubject
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByclassSubject/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByclassSubject
      tags:
        - Class
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
  /student:
    post:
      x-skaffolder-id: 5ff0ad09a3e049112323986d
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5fee9973a3e04911232394cb
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Student
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: object
                contact:
                  type: object
                dob:
                  type: integer
                email:
                  type: string
                firstName:
                  type: string
                gender:
                  type: string
                lastName:
                  type: string
    get:
      x-skaffolder-id: 5ff0ab0aa3e049112323985b
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5fee9973a3e04911232394cb
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Student
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
  "/student/{id}":
    delete:
      x-skaffolder-id: 5ff0ab0aa3e049112323985c
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5fee9973a3e04911232394cb
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ff0ad09a3e049112323986c
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5fee9973a3e04911232394cb
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    post:
      x-skaffolder-id: 5ff0ad09a3e049112323986e
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5fee9973a3e04911232394cb
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
  /subject:
    post:
      x-skaffolder-id: 5ff0aea3a3e049112323988f
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5ff09e9ea3e04911232397fd
      x-skaffolder-resource: Subject
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Subject
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subjectCode:
                  type: string
                subjectName:
                  type: string
    get:
      x-skaffolder-id: 5ff0ad42a3e0491123239878
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5ff09e9ea3e04911232397fd
      x-skaffolder-resource: Subject
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Subject
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
  "/subject/{id}":
    delete:
      x-skaffolder-id: 5ff0ae70a3e0491123239887
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ff09e9ea3e04911232397fd
      x-skaffolder-resource: Subject
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Subject
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ff0aea3a3e049112323988b
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ff09e9ea3e04911232397fd
      x-skaffolder-resource: Subject
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Subject
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    post:
      x-skaffolder-id: 5ff0aea3a3e049112323988c
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ff09e9ea3e04911232397fd
      x-skaffolder-resource: Subject
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Subject
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Subject"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Class:
      x-skaffolder-id: 5fee9a81a3e04911232394d1
      x-skaffolder-id-db: 5fee94e5a3e049112323945a
      x-skaffolder-id-entity: 5fee9a81a3e04911232394ce
      x-skaffolder-url: /class
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        classCode:
          type: string
          x-skaffolder-id-attr: 5fee9a81a3e04911232394cf
          x-skaffolder-type: String
          x-skaffolder-unique: true
        className:
          type: string
          x-skaffolder-id-attr: 5fee9a81a3e04911232394d0
          x-skaffolder-type: String
      x-skaffolder-relations:
        classSubject:
          x-skaffolder-id: 5ff0a982a3e0491123239858
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5fee9a81a3e04911232394ce
          x-skaffolder-ent2: 5ff09e9ea3e04911232397fa
      required:
        - _id
...
    Student:
      x-skaffolder-id: 5fee9973a3e04911232394cb
      x-skaffolder-id-db: 5fee94e5a3e049112323945a
      x-skaffolder-id-entity: 5fee9973a3e04911232394c5
      x-skaffolder-url: /student
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        address:
          type: object
          x-skaffolder-id-attr: 5ff0a55ba3e0491123239829
          x-skaffolder-type: Custom
        contact:
          type: object
          x-skaffolder-id-attr: 5fee9973a3e04911232394c9
          x-skaffolder-type: Custom
        dob:
          type: integer
          x-skaffolder-id-attr: 5fee9973a3e04911232394ca
          x-skaffolder-type: Date
        email:
          type: string
          x-skaffolder-id-attr: 5fee9973a3e04911232394c8
          x-skaffolder-type: String
        firstName:
          type: string
          x-skaffolder-id-attr: 5fee9973a3e04911232394c6
          x-skaffolder-type: String
        gender:
          type: string
          x-skaffolder-id-attr: 5ff0a5c9a3e0491123239842
          x-skaffolder-type: String
          x-skaffolder-enumeration:
            - FEMALE
            - MALE
        lastName:
          type: string
          x-skaffolder-id-attr: 5fee9973a3e04911232394c7
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
    Subject:
      x-skaffolder-id: 5ff09e9ea3e04911232397fd
      x-skaffolder-id-db: 5fee94e5a3e049112323945a
      x-skaffolder-id-entity: 5ff09e9ea3e04911232397fa
      x-skaffolder-url: /subject
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        subjectCode:
          type: string
          x-skaffolder-id-attr: 5ff09e9ea3e04911232397fc
          x-skaffolder-type: String
        subjectName:
          type: string
          x-skaffolder-id-attr: 5ff09e9ea3e04911232397fb
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
...